phone_tracker/requirements.txt
branca==0.4.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Flask==2.1.1
folium==0.12.1.post1
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
python-dotenv==0.20.0
requests==2.27.1
tinydb==4.7.0
urllib3==1.26.9
Werkzeug==2.1.1
wincertstore==0.2
zipp==3.8.0
